Fix/home pictures

See merge request dam973715/sprint-spirit!18

+++ app/src/main/java/com/example/sprintspirit/database/DBManager.kt
@@ -43,7 +43,7 @@ interface DBManager {
         email: String,
         password: String,
         onSuccess: () -> Unit,
-        onFailure: () -> Unit)
+        onFailure: (Int) -> Unit)
 
     fun signUpWithEmail(
         username: String,
